main.py
import streamlit as st
import pdfplumber
from PIL import Image
import pytesseract
import re
from textblob import TextBlob

# Function to convert PDF to text using pdfplumber
def pdf_to_text(file_path):
    with pdfplumber.open(file_path) as pdf:
        text = ''
        for page in pdf.pages:
            text += page.extract_text()
    return text

# Function to perform OCR on an image using Tesseract
def image_to_text(image_path):
    image = Image.open(image_path)
    text = pytesseract.image_to_string(image)
    return text

# Function for basic cleaning
def clean_text(text):
    """
    Clean the input text by removing headers, footers, special characters, etc.

    Parameters:
    - text (str): The input text to be cleaned.

    Returns:
    - cleaned_text (str): The cleaned text.
    """
    # 1. Remove Headers and Footers
    # Headers and footers often contain metadata or information that is not part of the main content.
    # Assuming headers and footers are located at the beginning and end of the text, respectively.

    # Define a pattern to match headers and footers
    header_footer_pattern = re.compile(r'^\s*[\w\s]+\s*$')
    
    # Remove headers
    text_lines = text.split('\n')
    text_lines = [line for line in text_lines if not header_footer_pattern.match(line)]

    # Remove footers
    text_lines = [line for line in reversed(text_lines) if not header_footer_pattern.match(line)]
    text_lines = list(reversed(text_lines))

    # Join the lines back into text
    text = '\n'.join(text_lines)

    # 2. Remove Special Characters and Encoding Issues
    # Special characters and encoding issues might interfere with downstream processing.

    # Remove non-alphanumeric characters except for whitespace
    text = re.sub(r'[^A-Za-z0-9\s]', '', text)

    # 3. Hyphenation and Line Breaks
    # Fix incorrect line breaks and hyphenated words split across lines.

    # Replace hyphenated words at line breaks
    text = re.sub(r'(\w+)-\s*\n\s*', r'\1', text)

    # 4. Handling Whitespace
    # Normalize whitespace, remove leading and trailing whitespaces
    text = ' '.join(text.split())

    # 5. Addressing Other Cleaning Tasks
    # Add more cleaning tasks as needed, such as handling specific characters or patterns unique to your data.

    # ...

    # Return the cleaned text
    return text


# Function for hyphenation and line breaks
def fix_line_breaks(text):
    """
    Fix line breaks and hyphenation issues in the input text.

    Parameters:
    - text (str): The input text with potential line break and hyphenation issues.

    Returns:
    - fixed_text (str): The text with corrected line breaks and hyphenation.
    """
    # 1. Fix Incorrect Line Breaks
    # Incorrect line breaks might split words or sentences unnaturally.

    # Merge lines that end with a hyphen
    text = re.sub(r'(\w)-\s*\n\s*', r'\1', text)

    # Merge lines that end with punctuation
    text = re.sub(r'([.!?])\s*\n\s*', r'\1 ', text)

    # 2. Handle Hyphenated Words Split Across Lines
    # Hyphenated words at the end of lines need to be joined.

    # Join hyphenated words at line breaks
    text = re.sub(r'(\w+)-\s*\n\s*', r'\1', text)

    # 3. Normalize Whitespace
    # Remove extra spaces and normalize whitespace

    # Replace multiple spaces with a single space
    text = ' '.join(text.split())

    # Return the text with corrected line breaks and hyphenation
    return text


# Function to normalize text
def normalize_text(text):
    """
    Normalize the input text by converting case and standardizing spelling.

    Parameters:
    - text (str): The input text to be normalized.

    Returns:
    - normalized_text (str): The normalized text.
    """
    # 1. Uniform Case Conversion
    # Convert the entire text to lowercase or uppercase, providing uniformity.

    # Convert text to lowercase
    text = text.lower()

    # 2. Standardize Spelling
    # Implement logic to standardize spelling, which may include correcting common errors.

    # Example: Replace common misspellings
    text = re.sub(r'\bteh\b', 'the', text)
    text = re.sub(r'\bu\b', 'you', text)

    # 3. Additional Spelling Standardization
    # Add more spelling standardization rules as needed for your specific use case.

    # ...

    # Return the normalized text
    return text


# Function to remove irrelevant content
def remove_irrelevant_content(text):
    """
    Remove non-essential sections, handle images, and tables in the input text.

    Parameters:
    - text (str): The input text containing potentially irrelevant content.

    Returns:
    - relevant_text (str): The text with irrelevant content removed.
    """
    # 1. Exclude Non-Essential Sections
    # Identify and remove sections that are typically non-essential, such as appendices or bibliographies.

    # Define patterns for non-essential sections
    non_essential_patterns = [
        re.compile(r'\bappendix\b', re.IGNORECASE),
        re.compile(r'\bbibliography\b', re.IGNORECASE),
        # Add more patterns as needed
    ]

    # Remove non-essential sections
    for pattern in non_essential_patterns:
        text = re.sub(pattern, '', text)

    # 2. Handle Images and Tables
    # Depending on the use case, you might want to transcribe text from images or tables, or you might choose to remove them entirely.

    # Example: Remove image captions
    text = re.sub(r'\[image: [^\]]+\]', '', text)

    # Example: Remove tables (assuming they are marked in some way)
    text = re.sub(r'\[table: [^\]]+\]', '', text)

    # 3. Additional Logic for Irrelevant Content
    # Add more logic to handle other types of irrelevant content based on your specific use case.

    # ...

    # Return the text with irrelevant content removed
    return text


# Function to correct errorsfrom textblob import TextBlob  # Make sure to install the 'textblob' package using: pip install textblob

def correct_errors(text):
    """
    Correct spelling and grammar errors in the input text.

    Parameters:
    - text (str): The input text with potential errors.

    Returns:
    - corrected_text (str): The text with corrected errors.
    """
    # 1. Spell-Check using TextBlob
    # TextBlob is a simple natural language processing library that includes a spell-check feature.
   
    # Create a TextBlob object for the input text
    blob = TextBlob(text)

    # Correct spelling errors
    corrected_text = str(blob.correct())

    # 2. Grammar Check (Optional)
    # TextBlob also provides a basic grammar-check feature. Uncomment the following lines if you want to perform grammar checking as well.

    # sentences = blob.sentences
    # for sentence in sentences:
    #     corrected_sentence = str(sentence.correct())
    #     corrected_text += corrected_sentence + ' '

    # 3. Additional Error Correction
    # Add more error correction logic based on your specific use case.

    # ...

    # Return the text with corrected errors
    return corrected_text


# Function for structural consistency
def ensure_consistency(text):
    """
    Ensure structural consistency in paragraphs and sentences in the input text.

    Parameters:
    - text (str): The input text with potential structural inconsistencies.

    Returns:
    - consistent_text (str): The text with improved structural consistency.
    """
    # 1. Ensure Proper Paragraph Structure
    # Identify and correct issues related to paragraph structure.

    # Define a pattern for identifying paragraphs
    paragraph_pattern = re.compile(r'\n\s*\n')

    # Ensure a single blank line between paragraphs
    text = re.sub(paragraph_pattern, '\n\n', text)

    # 2. Ensure Proper Sentence Structure
    # Address issues related to sentence structure.

    # Ensure a space after punctuation marks at the end of sentences
    text = re.sub(r'([.!?])\s*(?=[A-Z])', r'\1 ', text)

    # Ensure a single space between sentences
    text = re.sub(r'(?<=[.!?])\s*(?=[A-Z])', ' ', text)

    # 3. Additional Logic for Structural Consistency
    # Add more logic to handle other structural inconsistencies based on your specific use case.

    # ...

    # Return the text with improved structural consistency
    return text


# Function for data reductionfrom gensim.summarization import summarize  # Make sure to install the 'gensim' package using: pip install gensim

def reduce_data(text):
    """
    Reduce the amount of data by eliminating redundancies and summarizing lengthy sections.

    Parameters:
    - text (str): The input text with potential redundancies or lengthy sections.

    Returns:
    - reduced_text (str): The text with reduced data.
    """
    # 1. Eliminate Redundancies
    # Identify and eliminate redundant information.

    # Example: Remove consecutive duplicate words
    text = re.sub(r'\b(\w+)\s+\1\b', r'\1', text)

    # 2. Summarize Lengthy Sections
    # Use a text summarization tool to create concise summaries of lengthy sections.
    # Note: The example uses the gensim library for text summarization. Install it using 'pip install gensim'.

    # Summarize the text (Example: using Gensim's TextRank algorithm)
    #summarized_text = summarize(text)

    # 3. Additional Logic for Data Reduction
    # Add more logic to handle other types of data reduction based on your specific use case.

    # ...

    # Return the reduced text
    return text


# Function for final review and formatting
def final_review(text):
    """
    Perform a final review of the text, implement output formatting, and remove sensitive information.

    Parameters:
    - text (str): The input text to be reviewed and formatted.

    Returns:
    - formatted_text (str): The text after final review and formatting.
    """
    # 1. Remove Sensitive Information
    # Identify and remove any sensitive information that should not be included in the final output.

    # Example: Remove names (assuming names are sensitive information)
    #text = re.sub(r'\b[A-Z][a-z]+\b', '[REDACTED]', text)

    # 2. Output Formatting
    # Implement logic for final output formatting.

    # Example: Add a line break after each sentence
    text = re.sub(r'(?<=[.!?])\s*(?=[A-Z])', '\n', text)

    # Example: Ensure a consistent number of spaces after punctuation
    text = re.sub(r'([.,;:!?])\s*', r'\1 ', text)

    # 3. Additional Logic for Final Review and Formatting
    # Add more logic to handle other aspects of final review and formatting based on your specific use case.

    # ...

    # Return the formatted text
    return text


# Streamlit app
def main():
    st.title("PDF Text Processing App")

    uploaded_file = st.file_uploader("Upload a PDF file", type=["pdf"])

    if uploaded_file is not None:
        text = pdf_to_text(uploaded_file)

        st.subheader("Original Text:")
        st.text(text)

        # Perform text processing steps
        text = clean_text(text)
        text = fix_line_breaks(text)
        text = normalize_text(text)
        text = remove_irrelevant_content(text)
        text = correct_errors(text)
        text = ensure_consistency(text)
        text = reduce_data(text)
        text = final_review(text)

        st.subheader("Processed Text:")
        st.text(text)

if __name__ == "__main__":
    main()
